+++ src/main/kotlin/com/zepben/auth/server/JWTAuthenticator.kt
@@ -31,6 +31,7 @@ import java.security.interfaces.RSAPublicKey
 const val WELL_KNOWN_JWKS_PATH = "/.well-known/jwks.json"
 const val AUTHORIZATION_HEADER = "Authorization"
 const val CONTENT_TYPE = "Content-Type"
+const val METADATA = "Metadata"
 
 data class AuthResponse(
     val statusCode: StatusCode,